requirements.txt
torch --index-url https://download.pytorch.org/whl/cu118
torchvision --index-url https://download.pytorch.org/whl/cu118
transformers
streamlit
pandas
numpy
dvc
mlflow
scikit-learn
PyYAML
flask
flask-cors
langchain
langchain-groq
langchain-community 
google-search-results
serpapi
xmltodict